feat(dbus,outputs): merge OutputsService into State
State now handles DBus setup / registration when it is ready and each MetaHead now does its registration of metamodes internally rather than it being in main. In practice it is the same, just shifted where it actually happens.

Added some more signals to OutputManager so we can more reliably emit signals when global state has changed in Outputs. This should resolve #4.

Added a generic stateChanged signal on MetaHead to make connect / disconnection and waterfalling of signal to State easier to manage.